+++ scripts/setup-wake-on-lan.js
@@ -65,6 +65,7 @@ public class WakeOnLanModule extends ReactContextBaseJavaModule {
                             OkHttpClient client = OkHttpClientProvider.getOkHttpClient();
                             if (client != null) {
                                 client.connectionPool().evictAll();
+                                client.dispatcher().cancelAll();
                             }
                         } catch (Throwable ignored) {}
                     }
@@ -75,6 +76,7 @@ public class WakeOnLanModule extends ReactContextBaseJavaModule {
                             OkHttpClient client = OkHttpClientProvider.getOkHttpClient();
                             if (client != null) {
                                 client.connectionPool().evictAll();
+                                client.dispatcher().cancelAll();
                             }
                         } catch (Throwable ignored) {}
                     }